moved LearnFragment from .ui.main package
