+++ MyFitnessRoutines/src/main/java/com/devindow/myfitnessroutines/TabbedActivity.java
@@ -6,11 +6,7 @@ import android.support.design.widget.Snackbar;
 import android.support.design.widget.TabLayout;
 import android.support.v4.view.ViewPager;
 import android.support.v7.app.AppCompatActivity;
-import android.view.Menu;
-import android.view.MenuItem;
 import android.view.View;
-
-import com.devindow.myfitnessroutines.ui.main.SectionsPagerAdapter;
 
 public class TabbedActivity extends AppCompatActivity {
