+++ MyFitnessRoutines/src/main/res/layout/fragment_learn.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.devindow.myfitnessroutines.ui.main.LearnFragment">
+    tools:context="com.devindow.myfitnessroutines.LearnFragment">
 
     <Switch
         android:id="@+id/switch1"
